+++ plugin-lifecycle/src/main/java/com/mapbox/maps/plugin/lifecycle/MapboxLifecyclePluginImpl.kt
@@ -10,12 +10,8 @@ import android.content.ComponentCallbacks2.TRIM_MEMORY_RUNNING_MODERATE
 import android.content.ComponentCallbacks2.TRIM_MEMORY_UI_HIDDEN
 import android.content.res.Configuration
 import android.view.View
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleObserver
-import androidx.lifecycle.OnLifecycleEvent
-import androidx.lifecycle.ViewTreeLifecycleOwner
+import androidx.lifecycle.*
 import com.mapbox.maps.MapboxLifecycleObserver
-import com.mapbox.maps.logE
 import com.mapbox.maps.logW
 import com.mapbox.maps.plugin.Plugin.Companion.MAPBOX_LIFECYCLE_PLUGIN_ID
 import com.mapbox.maps.plugin.delegates.MapPluginProviderDelegate
@@ -31,37 +27,32 @@ class MapboxLifecyclePluginImpl : MapboxLifecyclePlugin {
    * @param observer the observer that listen to the life cycle events
    */
   override fun registerLifecycleObserver(mapView: View, observer: MapboxLifecycleObserver) {
-    val lifecycleOwner = ViewTreeLifecycleOwner.get(mapView)
-    if (lifecycleOwner == null) {
-      logE(
-        TAG,
-        """Can't get lifecycleOwner for mapview,
-          please make sure the host Activity is AppCompatActivity and the version of appcompat is 1.3.0+.
-          If the host Activity is not AppCompatActivity,
-          you need manually invoke the corresponding lifecycle methods in onStart/onStop/onDestroy/onLowMemory methods of the host Activity"""
-      )
-    } else {
-      val componentCallback = object : ComponentCallbacks2 {
-        override fun onConfigurationChanged(newConfig: Configuration) {
-          // no need
-        }
+    val viewLifecycleRegistry = ViewLifecycleOwner(
+      view = mapView
+    )
 
-        override fun onLowMemory() {
-          observer.onLowMemory()
-        }
+    val componentCallback = object : ComponentCallbacks2 {
+      override fun onConfigurationChanged(newConfig: Configuration) {
+        // no need
+      }
 
-        override fun onTrimMemory(level: Int) {
-          when (level) {
-            TRIM_MEMORY_RUNNING_CRITICAL, TRIM_MEMORY_RUNNING_LOW -> {
-              logW(TAG, "onTrimMemory with level $level is received, reduceMemoryUse will be called.")
-              observer.onLowMemory()
-            }
-            TRIM_MEMORY_BACKGROUND, TRIM_MEMORY_COMPLETE, TRIM_MEMORY_MODERATE, TRIM_MEMORY_RUNNING_MODERATE, TRIM_MEMORY_UI_HIDDEN -> Unit
+      override fun onLowMemory() {
+        observer.onLowMemory()
+      }
+
+      override fun onTrimMemory(level: Int) {
+        when (level) {
+          TRIM_MEMORY_RUNNING_CRITICAL, TRIM_MEMORY_RUNNING_LOW -> {
+            logW(TAG, "onTrimMemory with level $level is received, reduceMemoryUse will be called.")
+            observer.onLowMemory()
           }
+          TRIM_MEMORY_BACKGROUND, TRIM_MEMORY_COMPLETE, TRIM_MEMORY_MODERATE, TRIM_MEMORY_RUNNING_MODERATE, TRIM_MEMORY_UI_HIDDEN -> Unit
         }
       }
-      mapView.context.registerComponentCallbacks(componentCallback)
-      lifecycleOwner.lifecycle.addObserver(object : LifecycleObserver {
+    }
+    mapView.context.registerComponentCallbacks(componentCallback)
+    viewLifecycleRegistry.lifecycle.addObserver(
+      object : LifecycleObserver {
         @OnLifecycleEvent(Lifecycle.Event.ON_START)
         fun onStart() {
           observer.onStart()
@@ -75,11 +66,11 @@ class MapboxLifecyclePluginImpl : MapboxLifecyclePlugin {
         @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
         fun onDestroy() {
           observer.onDestroy()
-          lifecycleOwner.lifecycle.removeObserver(this)
+          viewLifecycleRegistry.lifecycle.removeObserver(this)
+          viewLifecycleRegistry.cleanUp()
           mapView.context.unregisterComponentCallbacks(componentCallback)
         }
       })
-    }
   }
 
   private companion object {